Rename pkg/console/subresource -> pkg/console/resource. Subresource has specific meaning in kube
