+++ hack/test-unit.sh
@@ -10,13 +10,13 @@ PACKAGES_TO_TEST=(
     "github.com/openshift/console-operator/pkg/cmd/version"
     "github.com/openshift/console-operator/pkg/console/operator"
     "github.com/openshift/console-operator/pkg/console/starter"
-    "github.com/openshift/console-operator/pkg/console/subresource/configmap"
-    "github.com/openshift/console-operator/pkg/console/subresource/deployment"
-    "github.com/openshift/console-operator/pkg/console/subresource/oauthclient"
-    "github.com/openshift/console-operator/pkg/console/subresource/route"
-    "github.com/openshift/console-operator/pkg/console/subresource/secret"
-    "github.com/openshift/console-operator/pkg/console/subresource/service"
-    "github.com/openshift/console-operator/pkg/console/subresource/util"
+    "github.com/openshift/console-operator/pkg/console/resource/configmap"
+    "github.com/openshift/console-operator/pkg/console/resource/deployment"
+    "github.com/openshift/console-operator/pkg/console/resource/oauthclient"
+    "github.com/openshift/console-operator/pkg/console/resource/route"
+    "github.com/openshift/console-operator/pkg/console/resource/secret"
+    "github.com/openshift/console-operator/pkg/console/resource/service"
+    "github.com/openshift/console-operator/pkg/console/resource/util"
     "github.com/openshift/console-operator/pkg/console/version"
     "github.com/openshift/console-operator/pkg/controller"
     "github.com/openshift/console-operator/pkg/crypto"
