+++ pkg/console/operator/operator.go
@@ -3,12 +3,12 @@ package operator
 import (
 	// standard lib
 	"fmt"
-	"github.com/openshift/console-operator/pkg/console/subresource/configmap"
-	"github.com/openshift/console-operator/pkg/console/subresource/deployment"
-	"github.com/openshift/console-operator/pkg/console/subresource/oauthclient"
-	"github.com/openshift/console-operator/pkg/console/subresource/route"
-	"github.com/openshift/console-operator/pkg/console/subresource/secret"
-	"github.com/openshift/console-operator/pkg/console/subresource/service"
+	"github.com/openshift/console-operator/pkg/console/resource/configmap"
+	"github.com/openshift/console-operator/pkg/console/resource/deployment"
+	"github.com/openshift/console-operator/pkg/console/resource/oauthclient"
+	"github.com/openshift/console-operator/pkg/console/resource/route"
+	"github.com/openshift/console-operator/pkg/console/resource/secret"
+	"github.com/openshift/console-operator/pkg/console/resource/service"
 
 	// 3rd party
 	"github.com/blang/semver"
