+++ pkg/console/resource/secret/secret.go
@@ -7,8 +7,8 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	// openshift
 	"github.com/openshift/console-operator/pkg/apis/console/v1alpha1"
-	"github.com/openshift/console-operator/pkg/console/subresource/deployment"
-	"github.com/openshift/console-operator/pkg/console/subresource/util"
+	"github.com/openshift/console-operator/pkg/console/resource/deployment"
+	"github.com/openshift/console-operator/pkg/console/resource/util"
 )
 
 const ClientSecretKey = "clientSecret"
